l3

+++ LibraryCPP/Tests/vector.cpp
@@ -69,7 +69,7 @@ int main()
         vector_set(vector, i - 1, i);
     }
 
-    long long sum = 0;
+    int sum = 0;
     for (int i = 0 ; i < 10000000 ; ++i)
         sum += vector_get(vector, i);
